+++ src/constants.js
@@ -1,0 +1,49 @@
+// Nishuuu Themes Constants
+
+const ACCENT_COLORS = {
+    'Lime': '#39EA5F',
+    'Blue': '#5393FF',
+    'Cyan': '#57D7FF',
+    'Bright Teal': '#64FFDA',
+    'Indigo': '#758AFF',
+    'Teal': '#80CBC4',
+    'Purple': '#B54DFF',
+    'Neon Lime': '#C6FF00',
+    'Vira': '#E9A581',
+    'Tomato': '#F85044',
+    'Pink': '#FF669E',
+    'Orange': '#FF7042',
+    'Yellow': '#FFCF3D',
+    'White': '#FFFFFF'
+};
+
+const FOLDER_TYPES = [
+    'folder_api_open', 'folder_app_open', 'folder_assets_open', 'folder_changesets_open',
+    'folder_ci_open', 'folder_components_open', 'folder_config_open', 'folder_controller_open',
+    'folder_coverage_open', 'folder_css_open', 'folder_database_open', 'folder_decorators_open',
+    'folder_dist_open', 'folder_docs_open', 'folder_fonts_open', 'folder_git_open',
+    'folder_github_open', 'folder_gitlab_open', 'folder_gulp_open', 'folder_husky_open',
+    'folder_images_open', 'folder_intellij_open', 'folder_jest_open', 'folder_js_open',
+    'folder_locales_open', 'folder_log_open', 'folder_middleware_open', 'folder_netlify_open',
+    'folder_next_open', 'folder_nginx_open', 'folder_node_open', 'folder_nuxt_open',
+    'folder_open', 'folder_packages_open', 'folder_plugin_open', 'folder_public_open',
+    'folder_root_open', 'folder_routes_open', 'folder_sass_open', 'folder_scripts_open',
+    'folder_server_open', 'folder_src_open', 'folder_storybook_open', 'folder_svg_open',
+    'folder_temp_open', 'folder_test_open', 'folder_turbo_open', 'folder_types_open',
+    'folder_utils_open', 'folder_vscode_open', 'folder_webpack_open', 'folder_wordpress_open'
+];
+
+const ICON_THEME_FILES = [
+    'tara-icons-carbon.json',
+    'tara-icons-deepforest.json',
+    'tara-icons-graphene.json',
+    'tara-icons-ocean.json',
+    'tara-icons-palenight.json',
+    'tara-icons-teal.json'
+];
+
+module.exports = {
+    ACCENT_COLORS,
+    FOLDER_TYPES,
+    ICON_THEME_FILES
+};